refactor: move snakeToCamel function to stringUtils and update imports
