+++ src/core/utils/conversion.ts
@@ -1,18 +1,4 @@
-/**
- * A private helper that converts a single snake_case or kebab-case string to camelCase.
- * e.g., 'user_profile' -> 'userProfile'
- */
-export function snakeToCamel(str: string): string {
-  var parts = str.split('_').filter((part) => part.length > 0);
-  if (parts.length === 0) return '';
-
-  return parts.reduce(function (p, c, index) {
-    if (index === 0) {
-      return c.charAt(0).toLowerCase() + c.slice(1);
-    }
-    return p + c.charAt(0).toUpperCase() + c.slice(1);
-  }, '');
-}
+import { snakeToCamel } from './stringUtils';
 
 /**
  * Recursively traverses an object or an array and converts all keys from
@@ -65,3 +51,6 @@ export function toUrlQueryParams(params: Record<string, any>): string {
   }
   return searchParams.toString();
 }
+
+// Re-export snakeToCamel for backward compatibility
+export { snakeToCamel };
